login, sign up, home screen navigation
Made log in, sign up, and home screens and set up the navigation for them
