+++ frontend/src/screens/types.tsx
@@ -1,0 +1,5 @@
+export type RootStackParamList = {
+    Home: undefined;
+    LogIn: undefined; 
+    SignUp: undefined;
+};
